+++ src/Google/AdsApi/Dfp/Util/v201702/ReportDownloader.php
@@ -16,11 +16,11 @@
  */
 namespace Google\AdsApi\Dfp\Util\v201702;
 
-use Google\AdsApi\Common\GuzzleLogMessageHandler;
+use Google\AdsApi\Common\AdsGuzzleHttpClientFactory;
+use Google\AdsApi\Common\GuzzleHttpClientFactory;
 use Google\AdsApi\Dfp\v201702\ReportJobStatus;
 use Google\AdsApi\Dfp\v201702\ReportService;
 use GuzzleHttp\Client;
-use GuzzleHttp\HandlerStack;
 use GuzzleHttp\RequestOptions;
 use UnexpectedValueException;
 
@@ -51,14 +51,17 @@ class ReportDownloader {
    * @param int $reportJobId
    * @param int|null $pollTimeSeconds optional, specify the time to sleep, in
    *     seconds, when polling for a report's tatus
-   * @param Client|null $httpClient optional, the Guzzle HTTP client that will
-   *     handle HTTP calls to the report service
+   * @param Client|null $httpClient optional, the Guzzle HTTP client whose
+   *     handler stacks this library's logging middleware will be pushed to
+   * @param GuzzleHttpClientFactory|null $httpClientFactory optional, the Guzzle
+   *     HTTP client factory that will generate a client handling HTTP calls
    */
   public function __construct(
       ReportService $reportService,
       $reportJobId,
       $pollTimeSeconds = null,
-      Client $httpClient = null
+      Client $httpClient = null,
+      GuzzleHttpClientFactory $httpClientFactory = null
   ) {
     $this->logger =
         $reportService->getAdsSession()->getReportDownloaderLogger();
@@ -67,14 +70,11 @@ class ReportDownloader {
     $this->pollTimeSeconds = $pollTimeSeconds === null
         ? self::DEFAULT_POLL_SECONDS : $pollTimeSeconds;
 
-    if ($httpClient === null) {
-      $stack = HandlerStack::create();
-      $stack->before(
-          'http_errors', GuzzleLogMessageHandler::log($this->logger));
-      $this->httpClient = new Client(['handler' => $stack]);
-    } else {
-      $this->httpClient = $httpClient;
+    if ($httpClientFactory === null) {
+      $httpClientFactory =
+          new AdsGuzzleHttpClientFactory($this->logger, $httpClient);
     }
+    $this->httpClient = $httpClientFactory->generateHttpClient();
   }
 
   /**