+++ tests/Google/AdsApi/AdWords/BatchJobs/BatchJobUploadStatusTest.php
@@ -16,10 +16,15 @@
  */
 namespace Google\AdsApi\AdWords\BatchJobs;
 
+use Google\AdsApi\AdWords\AdWordsSessionBuilder;
+use Google\AdsApi\AdWords\Testing\BatchJobs\SimpleGuzzleHttpClientFactory;
+use Google\Auth\FetchAuthTokenInterface;
 use GuzzleHttp\Client;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Psr7\Response;
+use Monolog\Handler\NullHandler;
+use Monolog\Logger;
 use PHPUnit_Framework_TestCase;
 
 /**
@@ -35,6 +40,27 @@ class BatchJobUploadStatusTest extends PHPUnit_Framework_TestCase {
       'https://www.googleapis.com/resume-upload';
   private static $BYTES_OF_256K = 262144;
 
+  private $adWordsSession;
+
+  /**
+   * @see PHPUnit_Framework_TestCase::setUp
+   */
+  protected function setUp() {
+    $fetchAuthTokenInterfaceStub = $this
+        ->getMockBuilder(FetchAuthTokenInterface::class)
+        ->disableOriginalConstructor()
+        ->getMock();
+    $fetchAuthTokenInterfaceStub
+        ->method('fetchAuthToken')
+        ->willReturn(['access_token' => 'abc123']);
+    $this->adWordsSession = (new AdWordsSessionBuilder())
+        ->withDeveloperToken('ABcdeFGH93KL-NOPQ_STUv')
+        ->withClientCustomerId('123-456-7890')
+        ->withUserAgent('batch jobs delegate')
+        ->withOAuth2Credential($fetchAuthTokenInterfaceStub)
+        ->withBatchJobsUtilLogger(new Logger('', [new NullHandler()]))
+        ->build();
+  }
   /**
    * @covers Google\AdsApi\AdWords\BatchJobs\BatchJobUploadStatus::__construct
    * @covers Google\AdsApi\AdWords\BatchJobs\BatchJobUploadStatus::initiateResumableUpload
@@ -47,8 +73,13 @@ class BatchJobUploadStatusTest extends PHPUnit_Framework_TestCase {
             ['Location' => [self::$DUMMY_RESUMABLE_UPLOAD_URL]])]);
     $handler = HandlerStack::create($mockHandler);
     $httpClient = new Client(['handler' => $handler]);
-    $batchJobUploadStatus =
-        new BatchJobUploadStatus(self::$DUMMY_UPLOAD_URL, 0, $httpClient);
+    $batchJobUploadStatus = new BatchJobUploadStatus(
+        self::$DUMMY_UPLOAD_URL,
+        $this->adWordsSession,
+        0,
+        $httpClient,
+        new SimpleGuzzleHttpClientFactory($httpClient)
+    );
     $this->assertSame(self::$DUMMY_RESUMABLE_UPLOAD_URL,
         $batchJobUploadStatus->getResumableUploadUrl());
     $this->assertSame(0, $batchJobUploadStatus->getTotalContentBytes());
@@ -58,9 +89,11 @@ class BatchJobUploadStatusTest extends PHPUnit_Framework_TestCase {
    * @covers Google\AdsApi\AdWords\BatchJobs\BatchJobUploadStatus::__construct
    */
   public function testCreateBatchJobUploadStatus_withNonZeroBytes() {
-    $batchJobUploadStatus =
-        new BatchJobUploadStatus(self::$DUMMY_UPLOAD_URL,
-            self::$BYTES_OF_256K);
+    $batchJobUploadStatus = new BatchJobUploadStatus(
+        self::$DUMMY_UPLOAD_URL,
+        $this->adWordsSession,
+        self::$BYTES_OF_256K
+    );
     $this->assertSame(self::$DUMMY_UPLOAD_URL,
         $batchJobUploadStatus->getResumableUploadUrl());
     $this->assertSame(self::$BYTES_OF_256K,